Moving augmentAttribute to utils.
